+++ src/asterix/InputParser.cpp
@@ -34,7 +34,7 @@ InputParser::InputParser(AsterixDefinition* pDefinition)
 /*
  * Parse data
  */
-AsterixData* InputParser::parsePacket(const unsigned char* m_pBuffer, unsigned int m_nBufferSize, unsigned long nTimestamp)
+AsterixData* InputParser::parsePacket(const unsigned char* m_pBuffer, unsigned int m_nBufferSize, double nTimestamp)
 {
   AsterixData* pAsterixData = new AsterixData();
   unsigned int m_nPos = 0;
@@ -112,7 +112,7 @@ AsterixData* InputParser::parsePacket(const unsigned char* m_pBuffer, unsigned i
 }
 
 DataBlock* InputParser::parse_next_data_block(const unsigned char* m_pData, unsigned int &m_nPos, unsigned int m_nBufferSize,
-    unsigned long nTimestamp, unsigned int &m_nDataLength)
+    double nTimestamp, unsigned int &m_nDataLength)
 /*  parse next data block
  *  AUTHOR: Krzysztof Rutkowski, ICM UW, krutk@icm.edu.pl
  */
